Introduce iflabel to name files
The names of config and service files is somewhat inconsistent,
some times using fff.batX and some times using batX.

This introduces an additional variable for the "label" used in
such paths.

In addition, this allows to call the functions with custom labels
or easy modification of those names if the entire script is used.

Finally, the introduction of the batif and iflabel variables
eliminates the repeated use of "fff.bat$bat" and "bat$bat",
which is somewhat annoying.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>